pyproject.toml
[project]
name = "2026-tft-japan-demo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "snowflake-connector-python>=4.3.0",
    "streamlit>=1.55.0",
]
